Use render FD for GBM allocator
We were using the same DRM FD for two different gbm_device objects. This
resulted in GEM handles being identical for a given buffer, but each
buffer instance not being aware of the other. When one gbm_bo was destroyed,
the GEM handle was closed, invalidating the other gbm_bo.

Closes: https://github.com/emersion/glider/issues/4